fix(BinaryReader): readLong would sometime return incorrect long

+++ src/main/java/marshmalliow/core/io/BinaryReader.java
@@ -245,8 +245,8 @@ public class BinaryReader {
 
 		return ((long)readBuffer[0] << 56) + ((long)(readBuffer[1] & 255) << 48)
 				+ ((long)(readBuffer[2] & 255) << 40) + ((long)(readBuffer[3] & 255) << 32)
-				+ ((readBuffer[4] & 255) << 24) + ((readBuffer[5] & 255) << 16)
-				+ ((readBuffer[6] & 255) << 8) + (readBuffer[7] & 255);
+				+ ((long)(readBuffer[4] & 255) << 24) + ((long)(readBuffer[5] & 255) << 16)
+				+ ((long)(readBuffer[6] & 255) << 8) + (long)(readBuffer[7] & 255);
 	}
 
 	/**
